+++ python/pounce/jax/__init__.py
@@ -44,7 +44,7 @@ _jax_config.update("jax_enable_x64", True)
 
 from ._build import from_jax
 from ._diff import solve, solve_with_warm, vmap_solve, vmap_solve_parallel
-from ._problem import JaxProblem
+from ._problem import AnchorState, JaxProblem
 
 __all__ = [
     "from_jax",
@@ -53,4 +53,5 @@ __all__ = [
     "vmap_solve",
     "vmap_solve_parallel",
     "JaxProblem",
+    "AnchorState",
 ]